requirements.txt

# Core - Async & Networking

asyncio==3.4.3
aiohttp==3.9.3
aioredis==2.0.1
zeromq==0.0.0          # via pyzmq
pyzmq==25.1.2


# Distributed State

redis==5.0.1


# Serialization

msgpack==1.0.7
pydantic==2.5.3


# Testing

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
locust==2.20.1          # load testing


# Monitoring & Metrics

prometheus-client==0.19.0
psutil==5.9.7


# Logging

structlog==24.1.0


# Utilities

python-dotenv==1.0.0
click==8.1.7            # CLI untuk menjalankan node
rich==13.7.0            # pretty terminal output